requirements.txt
# Build Tools (must be first for Python 3.12)
pip>=25.0
setuptools>=69.0.0
wheel>=0.42.0
packaging>=23.2

# Core Web Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.9

# Database & ORM
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
databases[postgresql]==0.9.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# AI/ML Libraries (CPU only)
transformers==4.46.3
torch==2.5.1
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
nltk==3.9.1
sentence-transformers==3.3.1
spacy==3.8.2

# Text Processing
textblob==0.18.0.post0

# HTTP & File Processing
httpx==0.28.1
aiofiles==24.1.0

# Validation & Data
pydantic==2.10.3
email-validator==2.2.0

# System Monitoring
psutil==6.1.0

# Development Tools
python-dotenv==1.0.1
loguru==0.7.2

# Testing (Optional)
pytest==8.3.3
pytest-asyncio==0.25.2

# Production (Optional)
gunicorn==23.0.0
redis==5.2.1
